+++ packages/contracts/test/systems/CounterTest.t.sol
@@ -19,7 +19,7 @@ contract CounterTest is PrimodiumTest {
 
   function testCounter() public {
     // Expect the counter to be 1 because it was incremented in the PostDeploy script.
-    uint32 counter = Counter.get(world);
+    uint256 counter = Counter.get(world);
     assertEq(counter, 1);
 
     // Expect the counter to be 2 after calling increment.
